+++ src/components/tool/texTitleRaw.tsx
@@ -1,115 +1,22 @@
 import { css } from "../../../styled-system/css";
-import { useRef, useEffect, useCallback, useState } from "react";
+import { useEffect } from "react";
+import {
+  parseBibTeX,
+  makeShorter,
+  convertToString,
+  useTexCommon,
+} from "./texCommon";
+import type { BibEntry } from "./texCommon";
 
 export const TexTitleRaw = () => {
-  const textAreaRef = useRef<HTMLTextAreaElement>(null);
-  const buttonRef = useRef<HTMLButtonElement>(null);
-  const resultRef = useRef<HTMLPreElement>(null);
-
-  interface BibEntry {
-    type: string;
-    id: string;
-    fields: Record<string, string>;
-  }
-
-  const parseBibTeX = (text: string) => {
-    const entries: BibEntry[] = [];
-    const entryRegex = /@(\w+)\s*{\s*([^,]*),([^@]*)}/g;
-
-    let match;
-    while ((match = entryRegex.exec(text)) !== null) {
-      const [, type, id, fieldsText] = match;
-      const fields: Record<string, string> = {};
-
-      // フィールドを1つずつ解析
-      let remaining = fieldsText.trim();
-      while (remaining) {
-        // キーを抽出
-        const keyMatch = /^\s*(\w+)\s*=\s*/.exec(remaining);
-        if (!keyMatch) break;
-
-        const key = keyMatch[1].toLowerCase();
-        remaining = remaining.slice(keyMatch[0].length);
-
-        // 値を抽出（中括弧のネストを考慮）
-        let value = "";
-        let braceCount = 0;
-        let i = 0;
-
-        if (remaining[0] === "{") {
-          braceCount = 1;
-          i = 1;
-          while (i < remaining.length && braceCount > 0) {
-            if (remaining[i] === "{") braceCount++;
-            else if (remaining[i] === "}") braceCount--;
-            if (braceCount > 0) value += remaining[i];
-            i++;
-          }
-        } else if (remaining[0] === '"') {
-          i = 1;
-          while (i < remaining.length && remaining[i] !== '"') {
-            value += remaining[i];
-            i++;
-          }
-          i++; // 終わりの " をスキップ
-        }
-
-        if (key && value) {
-          fields[key] = value;
-        }
-
-        // 次のフィールドの開始位置まで進める
-        remaining = remaining.slice(i).replace(/^\s*,\s*/, "");
-      }
-
-      entries.push({
-        type: type.toLowerCase(),
-        id,
-        fields,
-      });
-    }
-    console.log(entries);
-    return entries;
-  };
-
-  const makeShorter = (entries: BibEntry[]) => {
-    entries.forEach((entry) => {
-      // pagesのハイフンを直す
-      if (entry.fields.pages) {
-        entry.fields.pages = entry.fields.pages.replace("–", "-");
-        entry.fields.pages = entry.fields.pages.replace(/-+/g, "--");
-      }
-    });
-    return entries;
-  };
-
-  const convertToString = (entries: BibEntry[]) => {
-    return entries
-      .map((entry) => {
-        const regex = /(?<=( title)|( TITLE))\s*=\s*\{([^{}]+)\}/g;
-        const fields = Object.entries(entry.fields)
-          .map(([key, value]) => `  ${key} = {${value}}`)
-          .join(",\n")
-          .replace(regex, (match, p1, p2, p3) => ` = {{${p3}}}`);
-
-        return `@${entry.type}{${entry.id},\n${fields}\n}`;
-      })
-      .join("\n\n");
-  };
-
-  const [copyStatus, setCopyStatus] = useState<"default" | "copied">("default");
-
-  const handlePaste = useCallback(() => {
-    const textArea = textAreaRef.current;
-    const resultArea = resultRef.current;
-    if (textArea && resultArea) {
-      const bibTexContent = textArea.value;
-      const entries = parseBibTeX(bibTexContent);
-      const shortedEntries = makeShorter(entries);
-      const result = convertToString(shortedEntries);
-      resultArea.textContent = result;
-    }
-  }, []);
+  const {
+    textAreaRef,
+    buttonRef,
+    resultRef,
+    copyStatus,
+    setCopyStatus,
+    handlePaste,
+  } = useTexCommon();
 
   useEffect(() => {
     const button = buttonRef.current;
